+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `Attestation.sign` now only returns `AttestationError` when failing to sign a distribution file ([#XX]())
+
 ## [0.0.6]
 
 ### Added